+++ wled00/FX_fcn.cpp
@@ -199,6 +199,9 @@ void Segment::resetIfRequired() {
   if (data && _dataLen > 0) memset(data, 0, _dataLen);  // prevent heap fragmentation (just erase buffer instead of deallocateData())
   next_time = 0; step = 0; call = 0; aux0 = 0; aux1 = 0;
   reset = false;
+  #ifndef WLED_DISABLE_GIF
+  endImagePlayback(this);
+  #endif
 }
 
 CRGBPalette16 IRAM_ATTR &Segment::loadPalette(CRGBPalette16 &targetPalette, uint8_t pal) {