+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## WLED changelog
 
+#### Build 2403191
+
+-   Add Image effect (GIF playback from FS support)
+
 #### Build 2403190
 -   limit max PWM frequency (fix incorrect PWM resolution)
 -   Segment UI bugfix